+++ client/views/Facility/FacilityViewEdit.jsx
@@ -63,7 +63,8 @@ FacilityViewEdit = React.createClass({
 		}
 		return (
 		    <div className="ibox-form user-profile-card" style={{backgroundColor:"#fff"}}>
-			    <h2><span>{facility.getName()}</span></h2>
+			    <h2 style={{marginTop:"0px"}}>Edit facility</h2>
+			    {/*
 				<CollapseBox title="Property Details">
 					<div className="row">
 						<div className="col-sm-7">
@@ -98,6 +99,65 @@ FacilityViewEdit = React.createClass({
 				<CollapseBox title="Default services & suppliers" collapsed={true}>
 					<ServicesSelector item={facility} field={"servicesRequired"}/>
 				</CollapseBox>
+				*/}
+                <Stepper tabs={[
+                    {
+                        tab:<span id="discussion-tab">Basic Details</span>,
+                        content:
+                            <div className="row">
+								<div className="col-sm-7">
+									<AutoForm item={facility} schema={schema} form={["name","address","operatingTimes"]}/>
+								</div>
+					        	<div className="col-sm-5">
+					        		<DocThumb.File item={facility.thumb} onChange={facility.setThumb.bind(facility)} />
+					        	</div>
+			        		</div>,
+			        	instructions:<div>
+			        		Enter the basic facility info here including name, address, and image.
+			        	</div>
+                    },
+                    {
+                        tab:<span id="documents-tab">Documents</span>,
+                        content:
+                        	<AutoForm item={facility} schema={Facilities.schema()} form={["documents"]}/>,
+			        	instructions:<div>
+			        		Formal documentation related to the facility can be added here. This typically includes insurance and/or lease documents.
+			        	</div>
+                    },
+                    {
+                        tab:<span id="personnel-tab">Personnel</span>,
+                        content:
+                            <ContactList 
+                                items={members}
+                                facility={facility}
+                                onAdd={team&&team.canInviteMember()?this.addMember.bind(null,{role:"staff"}):null}
+                            />,
+			        	instructions:<div>
+			        		Enter the facility personnel here by clicking on add member.
+			        	</div>
+                    },
+                    {
+                        tab:<span id="suppliers-tab"><span>Suppliers</span></span>,
+                        content:
+                        	<ContactList 
+                                items={suppliers}
+                                facility={facility}
+                                type="team"
+                                onAdd={team&&team.canInviteSupplier()?this.addSupplier.bind(null,{role:"supplier"}):null}
+                            />,
+			        	instructions:<div>
+			        		Add the suppliers employed by this facility here. If you want to configure this later simply press next.
+			        	</div>
+                    },
+                    {
+                        tab:<span id="services-tab">Services</span>,
+                        content:
+                        	<ServicesSelector item={facility} field={"servicesRequired"}/>,
+			        	instructions:<div>
+			        		Enter the services required by this facility. If you want you can also match there services to a supplier. If you want to configure this later simply his finish.
+			        	</div>
+                    }
+                ]} />				
 			</div>
 		)
 	}